setup tests
